extract member component, delete unnecessary div
